# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: NexusID API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /bounties:
    get:
      operationId: listBounties
      tags: [bounties]
      summary: List bounties
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [open, claimed, closed]
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Bounties list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BountiesList"

    post:
      operationId: createBounty
      tags: [bounties]
      summary: Create a bounty
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBountyBody"
      responses:
        "201":
          description: Created bounty
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bounty"
  /bounties/{bountyId}:
    get:
      operationId: getBounty
      tags: [bounties]
      summary: Get a single bounty
      parameters:
        - name: bountyId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Bounty
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bounty"
        "404":
          description: Not found

    put:
      operationId: updateBounty
      tags: [bounties]
      summary: Update a bounty (poster only, open status only)
      parameters:
        - name: bountyId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBountyBody"
      responses:
        "200":
          description: Updated bounty
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bounty"
        "403":
          description: Unauthorized or bounty not open
        "404":
          description: Not found

    delete:
      operationId: deleteBounty
      tags: [bounties]
      summary: Delete a bounty (poster only, open status only)
      parameters:
        - name: bountyId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"
        "403":
          description: Unauthorized or bounty not open
        "404":
          description: Not found
  /bounties/{bountyId}/submissions:
    get:
      operationId: getBountySubmissions
      tags: [bounties]
      summary: List submissions for a bounty
      parameters:
        - name: bountyId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Submissions list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BountySubmissionsList"

    post:
      operationId: createBountySubmission
      tags: [bounties]
      summary: Submit a solution to a bounty
      parameters:
        - name: bountyId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBountySubmissionBody"
      responses:
        "201":
          description: Created submission
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BountySubmission"
  /bounties/{bountyId}/submissions/{submissionId}/winner:
    put:
      operationId: selectBountyWinner
      tags: [bounties]
      summary: Mark a submission as winner
      parameters:
        - name: bountyId
          in: path
          required: true
          schema:
            type: integer
        - name: submissionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Winner selected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BountySubmission"
  /bounties/{bountyId}/close:
    post:
      operationId: closeBounty
      tags: [bounties]
      summary: Close an open bounty (poster only)
      parameters:
        - name: bountyId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Bounty closed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bounty"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found

  # ─── Inner Circles ────────────────────────────────────────────────────────────
...
  /chat/screenshot:
    post:
      operationId: reportChatScreenshot
      tags: [chat]
      summary: Report that the current user took a screenshot of a DM thread
      description: |
        Mobile-only. Notifies the peer of the conversation and increments
        a public counter on the screenshotter's profile so others can see
        how often they screenshot DMs.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportChatScreenshotBody"
      responses:
        "200":
          description: Recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportChatScreenshotResponse"
        "401":
          description: Unauthorized
        "403":
          description: Not a participant of this conversation
        "404":
          description: Conversation not found
components:
  schemas:
    ReportChatScreenshotBody:
      type: object
      properties:
        conversationId:
          type: integer
        platform:
          type: string
          enum: [ios, android]
      required: [conversationId]
    ReportChatScreenshotResponse:
      type: object
      properties:
        ok:
          type: boolean
        chatScreenshotsTaken:
          type: integer
      required: [ok, chatScreenshotsTaken]
...
    UserProfile:
      type: object
      properties:
        id:
          type: string
        clerkId:
          type: string
        username:
          type: string
        displayName:
          type: string
        bio:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        coverUrl:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
        website:
          type: string
          nullable: true
        occupation:
          type: string
          nullable: true
        gender:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        skills:
          type: array
          items:
            type: string
        experience:
          type: array
          items:
            $ref: "#/components/schemas/Experience"
        didDocument:
          type: object
          nullable: true
          additionalProperties: true
        followersCount:
          type: integer
        followingCount:
          type: integer
        postsCount:
          type: integer
        ghostMode:
          type: boolean
        isFollowing:
          type: boolean
        bountiesWon:
          type: integer
        createdAt:
          type: string
        usernameChangedAt:
          type: string
          nullable: true
        lastSeenAt:
          type: string
          nullable: true
          description: Last time this user pinged the server. Powers the online dot.
        powerScore:
          type: ["integer", "null"]
        powerRank:
          type: ["string", "null"]
        isAdmin:
          type: boolean
          description: True when this user is a moderator/admin.
        verificationTier:
          type: string
          nullable: true
          enum: [silver, blue, null]
          description: |
            Identity-verification badge shown next to the display name.
            "silver" = identity verified, "blue" = notable / premium account,
            null = unverified. Granted by an admin via
            PATCH /admin/users/:userId/verification.
        chatScreenshotsTaken:
          type: integer
          description: |
            Public counter of how many times this user has screenshotted a
            DM thread (mobile only). Surfaced as a warning chip on the
            profile so peers can decide whether to share sensitive content.
      required: [id, clerkId, username, displayName, skills, experience, followersCount, followingCount, postsCount, ghostMode, isFollowing, bountiesWon, createdAt]
...
    Experience:
      type: object
      properties:
        title:
          type: string
        company:
          type: string
        startDate:
          type: string
        endDate:
          type: string
          nullable: true
        current:
          type: boolean
        description:
          type: string
          nullable: true
      required: [title, company, startDate, current]
...
    DeleteResult:
      type: object
      properties:
        success:
          type: boolean
      required: [success]
...
    Bounty:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        category:
          type: string
        reward:
          type: string
        deadline:
          type: string
          nullable: true
        status:
          type: string
          enum: [open, claimed, closed]
        poster:
          $ref: "#/components/schemas/UserProfile"
        submissionsCount:
          type: integer
        createdAt:
          type: string
      required: [id, title, description, category, reward, status, poster, submissionsCount, createdAt]
    BountiesList:
      type: object
      properties:
        bounties:
          type: array
          items:
            $ref: "#/components/schemas/Bounty"
        total:
          type: integer
      required: [bounties, total]
    CreateBountyBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
        reward:
          type: string
        deadline:
          type: string
      required: [title, description, category, reward]
    BountySubmission:
      type: object
      properties:
        id:
          type: integer
        bountyId:
          type: integer
        content:
          type: string
        link:
          type: string
          nullable: true
        isWinner:
          type: boolean
        submitter:
          $ref: "#/components/schemas/UserProfile"
        createdAt:
          type: string
      required: [id, bountyId, content, isWinner, submitter, createdAt]
    BountySubmissionsList:
      type: object
      properties:
        submissions:
          type: array
          items:
            $ref: "#/components/schemas/BountySubmission"
      required: [submissions]
    UpdateBountyBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
        reward:
          type: string
        deadline:
          type: string
          nullable: true
    CreateBountySubmissionBody:
      type: object
      properties:
        content:
          type: string
        link:
          type: string
      required: [content]

    # ─── Inner Circles ─────────────────────────────────────────────────────────
...